python-ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
mediapipe==0.10.8
tensorflow==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
python-multipart==0.0.6
pydantic==2.5.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pillow==10.1.0
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2